Large File Support
*Features Added:

Added large file support. User is able to edit or create files that are larger the view row limit. The user is able to scroll down the file with just using the down/up arrow keys. The user is also able to edit text that is past the view row limit.

*Known Issues:
Once the user scrolls down past the view row limit, there is always one row that displays empty; however, the user just needs to keep scrolling to see the remaining text.

Having wrapped text bugs the display out.